+++ frontend/apps/react-essentials-app/src/components/molecules/CardList.jsx
@@ -1,4 +1,4 @@
-import '@prj--reactjs-component-lib/generic-lsg/lib/molecules/cardList.css';
+import '@prj--reactjs-application-library/generic-lsg/lib/molecules/cardList.css';
 
 const CardList = ({ children, className }) => {
     return <ul className={`cardList ${className}`}>{children}</ul>;