+++ tools/build-scripts/DashboardGenerator.js
@@ -3,7 +3,7 @@ const path = require('path');
 
 const source = path.join(__dirname, '../..', 'package', 'apps');
 const dest = path.join(__dirname, '../..', 'package', 'apps');
-const BASE_URL = '/prj--reactjs-component-lib/';
+const BASE_URL = '/prj--reactjs-application-library/';
 
 const init = () => {
     // Read the contents of the directory