+++ frontend/components/_boilerplate-lib/CHANGELOG.md
@@ -3,24 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<<<<<<< Updated upstream
+
 ## [0.3.2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-lib@0.3.1...@prj--reactjs-component-lib/boilerplate-lib@0.3.2) (2023-11-08)
 
-**Note:** Version bump only for package @prj--reactjs-component-lib/boilerplate-lib
+# **Note:** Version bump only for package @prj--reactjs-component-lib/boilerplate-lib
+
+<<<<<<< Updated upstream
+
+> > > > > > > Stashed changes
 
 ## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-lib@0.3.0...@prj--reactjs-component-lib/boilerplate-lib@0.3.1) (2023-10-31)
 
-**Note:** Version bump only for package @prj--reactjs-component-lib/boilerplate-lib
+=======
 
-# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-lib@0.2.0...@prj--reactjs-component-lib/boilerplate-lib@0.3.0) (2023-10-30)
+## [0.3.2](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/boilerplate-lib@0.3.1...@prj--reactjs-application-library/boilerplate-lib@0.3.2) (2023-11-08)
+
+**Note:** Version bump only for package @prj--reactjs-application-library/boilerplate-lib
+
+## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/boilerplate-lib@0.3.0...@prj--reactjs-application-library/boilerplate-lib@0.3.1) (2023-10-31)
+
+> > > > > > > Stashed changes
+
+**Note:** Version bump only for package @prj--reactjs-application-library/boilerplate-lib
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/boilerplate-lib@0.2.0...@prj--reactjs-application-library/boilerplate-lib@0.3.0) (2023-10-30)
 
 ### Features
 
--   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
 
 # 0.2.0 (2023-10-26)
 
 ### Features
 
--   added some more componenst ([038f67e](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/038f67e70a49d759d0cefca505eb721ff9e6220e))
--   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
--   migrated concept cards ([c4170d2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/c4170d2130e71d04e587acd0f9a4f1becef4d0b3))
+-   added some more componenst ([038f67e](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/038f67e70a49d759d0cefca505eb721ff9e6220e))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   migrated concept cards ([c4170d2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/c4170d2130e71d04e587acd0f9a4f1becef4d0b3))