+++ frontend/design-systems/_boilerplate-lsg/CHANGELOG.md
@@ -3,27 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<<<<<<< Updated upstream
+
 # [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-lsg@0.3.1...@prj--reactjs-component-lib/boilerplate-lsg@0.4.0) (2023-11-08)
 
 ### Features
 
--   cleanup ([bdc0a92](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/bdc0a92044bd5b3691ec5f56c9ea2f86e3a7802e))
+-   # cleanup ([bdc0a92](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/bdc0a92044bd5b3691ec5f56c9ea2f86e3a7802e))
+    <<<<<<< Updated upstream
+    > > > > > > > Stashed changes
 
 ## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-lsg@0.3.0...@prj--reactjs-component-lib/boilerplate-lsg@0.3.1) (2023-10-31)
 
-**Note:** Version bump only for package @prj--reactjs-component-lib/boilerplate-lsg
+=======
 
-# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-component-lib/compare/@prj--reactjs-component-lib/boilerplate-lsg@0.2.0...@prj--reactjs-component-lib/boilerplate-lsg@0.3.0) (2023-10-30)
+# [0.4.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/boilerplate-lsg@0.3.1...@prj--reactjs-application-library/boilerplate-lsg@0.4.0) (2023-11-08)
 
 ### Features
 
--   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/f9780c2896d185c8adf83f5af0782939e799b430))
+-   cleanup ([bdc0a92](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/bdc0a92044bd5b3691ec5f56c9ea2f86e3a7802e))
+
+## [0.3.1](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/boilerplate-lsg@0.3.0...@prj--reactjs-application-library/boilerplate-lsg@0.3.1) (2023-10-31)
+
+> > > > > > > Stashed changes
+
+**Note:** Version bump only for package @prj--reactjs-application-library/boilerplate-lsg
+
+# [0.3.0](https://github.com/paulAlexSerban/prj--reactjs-application-library/compare/@prj--reactjs-application-library/boilerplate-lsg@0.2.0...@prj--reactjs-application-library/boilerplate-lsg@0.3.0) (2023-10-30)
+
+### Features
+
+-   **package scripts:** impl ([f9780c2](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/f9780c2896d185c8adf83f5af0782939e799b430))
 
 # 0.2.0 (2023-10-26)
 
 ### Features
 
--   added some more componenst ([038f67e](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/038f67e70a49d759d0cefca505eb721ff9e6220e))
--   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
--   remove unused file ([30e842f](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/30e842f323197eaf930f1edff58aa2b25eb5fc88))
--   workflow implementation ([dbae96d](https://github.com/paulAlexSerban/prj--reactjs-component-lib/commit/dbae96dfe108f8a2638051cb727efc6b86b606d4))
+-   added some more componenst ([038f67e](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/038f67e70a49d759d0cefca505eb721ff9e6220e))
+-   json ([ff6c1a7](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/ff6c1a7c419f4e66511235803ec26a9db5a85314))
+-   remove unused file ([30e842f](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/30e842f323197eaf930f1edff58aa2b25eb5fc88))
+-   workflow implementation ([dbae96d](https://github.com/paulAlexSerban/prj--reactjs-application-library/commit/dbae96dfe108f8a2638051cb727efc6b86b606d4))